"5:20pm. Holy shit. I just wanted to put in a few words and then get back to bed, but it seems it is already time to close.
I needed some time to reflect on my goals. Studying the libraries in the manner that I've wanted to is basically impossible.

Somebody with perfect control over his own desire could do it, but I sure as hell can't. 40k LOC in Coquelicot and who knows how much in the Reals library is just too much. I am already bored to tears by what I had.

5:25pm. Luckily, it does not seem like I am missing as much as I thought I would. I thought I needed some great insights in order to figure out the reals, but the proofs so far, while I could not easily produce them myself do not look anything out of the ordinary. I needed to learn some new things in order to absorb them, but not much beyond what I already knew.

I've decided to pick a new goal.

I want to go back to those Gaussian integral proofs and seeing how they could be replicated in Coq. That is what I wanted to do since the first time I saw them. It seems like a fitting challenge.

I operate much better when I am studying for the sake of something, rather than for the sake of studying itself.

Before I can even look at that though, I am going to have to check out how differentiation and integration proofs are done. The same goes for trigonometric functions.

5:30pm. I've already spent large amounts of time learning proofs. I am hardly a beginner at this point.

I should just gather my courage and chase after what I want directly which is statistics and linear algebra.

So for the next few days or so, I will continue my dabbling, but after I've gotten basic familiarity with Reals and Coquelicot I won't waste time boring myself to dead with an all encompassing study. I will instead pick a target and start peeling at it. That should be much more motivating than what I am doing now.

This is how I learned programming and this is how math should be learned as well. Studying a library from the ground up is sheer insanity.

Having all the proofs there should really help quite a lot with the peeling. Of course, I won't try to make the proof myself. Rather for such a basic thing no doubt somebody already did it. There should be a library of statistics proofs out there. That is what I need to study.

5:40pm. In summary, I am going to switch to high gear and change my strategy.

I'll do the same thing with the game I am playing. Forget finishing it, I think I'll just move straight into Long War.

5:45pm. Life is about doing things. Like I have nothing better to do than just study random stuff for the next few years. I already have a bunch of goals of what I want to do after I've attained math proficiency. I am completely clueless of what the implementation for the AI system will be, but I already understand how current formalisms are holding back its creation.

I have a very basic goal of porting primitive Bayesian inference (for conjugate models) to resemble transport over paths. If I can do that on the small, then that will open the door to building true hierarchies and to figuring out things like credit assignment and to bridging the model sampling and optimization perspectives. From the viewpoint of learning, equivalence is essentially a kind of computational reversibility. I am absolutely sure that this will play a large role in making credit assignment viable, but I am clueless how.

Most of the theoretical understanding is currently concentrated in flat models, but it has not been done correctly. If it were, then efficient hierarchies would not be so much of an issue.

If I can integrate that understanding in a way that satisfies me, it will be like turning on the lights. I will be able to crashing in the furniture and will be able to find my way around. It will be like climbing a tower in the midst of the thick forest in which everybody is lost in. I will be able to peer beyond the immediate and see where I really should go."