"11:10am. Saizan: "Basically transport (\ i -> H ?) u will get stuck if H is a indexed inductive type (higher or not), otherwise they still behave as in plain Agda or Coq."
That answers that. Yeah, they are going to have to make it work in order to complete the language.

And I haven't gotten a reply to my Agda question which confirms my suspicions about the state of the libraries. Though maybe I need to give it a bit more time.

11:20am. Today I got up late, because you guessed it - I stayed up finishing the XCOM 2 campaign. It was not like day 3 where the game had to crash for me to realize how late it really is. I made a conscious decision to keep playing and I do not regret it. And now it is out of my system. Unless I find something else, I won't stay up so late from here on out.

11:35am. The more I think about path concentration, the more I become convinced that it is really the only viable alternative to the competing frameworks. Bayesian inference and DL offers literally nothing in that regard. It is really an old idea brought to life by homotopy type theory. This hunch is the one that I need to chase.

That `merge_sort = insertion_sort` holds in Coq is an interesting observation, but higher inductive types crack open a wast space for optimization. There will be something to be something waiting to be discovered in that void.

11:45am. Right now I am thinking of how would I handle things at the top level. It always felt like the top would need a complete trace in order to propagate the credit, but why would that be? If the intermediate hierarchies have memory, then the top would only need the recent trace rather than the lifetime trace.

It could be made to work. It is just impossible otherwise. If it were impossible then the human brain would need the complete trace as well.

11:50am. But if that is true...then the hierarchies would need to be recurrent in depth as well as time. This is actually he case in the brain.

Even though the hints are everywhere, it is going to be really difficult finding the principle. But I was a guy who spent 1.5 decade thinking of a way to connect omnipotence to a real life process and finally found it. As hard as it will be it is not like there is a lack of intelligence to study in the real world. It will get done eventually.

11:55am. I do not want to misplace my hope again. But I am persistence personified. I am going to do this.

12pm. Since it is this late, I might as well take a break and then when I get back, I will dive back into the PRML book.

I've left it waiting for a while."