"11:40am. https://pron.github.io/posts/computation-logic-algebra-pt1
And now rather than do what I said, this caught my attention. Ah well, I'll do what I want today. I am done with late night gaming sessions so now I will be able to focus on things that are nearer to my real interest.

"This book-length text — and not a short book — is published as a three-part blog post series"

It will take a while it seems. Let me have breakfast here instead.

1:30pm. "Babbage was Equations From God, p. 153  “possibly the best-known academic mathematician who spoke his mind on theological issues… Babbage’s Ninth Bridgewater Treatise (1837) brazenly employed mathematical principles to support religious doctrine. For instance, he used equations from Laplace’s Théorie Analytique des Probabilités and Poisson’s Recherches sur la Probabilité des Jugements to counter David Hume’s repudiation of miracles, calculating the probability that so many witnesses could have been mistaken about seeing Jesus rise from the dead.”"

I've seen similar arguments to support the authenticity of climate change. Eric Drexler of Engines of Creation comes to mind.

And I've seen similar arguments employed to support the Fermi Paradox.

Once you are able to pull figures from your ass, you can find a reason to believe in anything.

2:20pm. https://pron.github.io/posts/computation-logic-algebra-pt2

I am in the middle of this currently. It is quite interesting. Religious undertones are abound in mathematical research.

"In other words, while it is intuitively reasonable to use 0 to represent the “empty” or “nothing” (“bottom”, in modern terms) element, there were, as we’ll see, different and possibly more intuitive proposals for the representation of the “universe” or “everything” (“top” in modern terms) element. But Boole liked the representation of 1, as it stand for the unity and totality of God (there are mathematical reasons, as well)"

2:45pm. "The motion of the hand performs the office of the mind."

I will remember this quote.

3:10pm. Let me take a break. I'll continue part 3 at a late date.

4:20pm. I've resumed. I really needed that bath.

https://pron.github.io/posts/tlaplus_part1
"There have also been some famous cases of algorithms that were ostensibly “proven”, but not formally — i.e., not in a formal system — that also turned out to be completely wrong, as discovered with the use of model checkers."

https://brooker.co.za/blog/2014/03/08/model-checking.html
"Good journals, well-known authors and informal proofs are not sufficient.

Mars Code, in February's CACM, is a very interesting look from the outside at some of the software engineering practices that helped make the Mars Science Laboratory mission successful. The authors cover a lot of ground in the article, from code reviews to coding standards to model extraction and model checking. While writing about model checking, they tell the story of Snark, a non-blocking deque algorithm.

The Snark paper looks great. The algorithm is presented clearly, including both clear text descriptions and blocks of C-like pseudocode. It's published in a well-respected journal, LNCS. It dedicates four and a half of its fifteen pages to a well-structured and clearly written sketch proof of the correctness of the algorithm, with clear diagrams explaining some of the tricky cases. It's got Guy Steele on the author list. It's from Sun. As far as my is this paper likely to be trustworthy? heuristics go, this one doesn't raise many suspicions.

Unfortunately, Snark is broken. Not subtly broken on an obscure liveness measure, but fundamentally broken in that it's unsafe. In DCAS is not a Silver Bullet for Nonblocking Algorithm Design, Doherty et al lay out two bugs in the Snark algorithm. It's worth noting, though, that et al in this case includes most of the authors of the original paper. The paper explains the bugs well, and it's interesting how subtly 30 lines of code can be broken. It's well worth reading, especially if you are interested in non-blocking algorithms. Later, Leslie Lamport used the same bugs as a test of the PlusCal language. In Checking a Multithreaded Algorithm with +CAL he explains how he model checked the original and fixed Snark algorithms using TLC."

Maybe it would be interesting to spend some time learning this. I'll read the first TLA post and then decide. Ron is a pretty good writer.

5:10pm. The first post was pretty hype. I am definitely interested in learning more about TLA.

But I had enough for today.

Somehow I just want to run away from this all, but to do that I'd literally have to run to a different universe. In order to get excited about what I am doing, I need to get excited about doing that single problem. I need to focus it my all on that single rewrite.

The problem is not that, but the immense and poorly documented libraries that underpin the whole thing. That is the real hard part of it all.

One thing I would enjoy at this time is some kind of course that covers the math in the Reals and the Coquelicot library.

5:20pm. But I will have to make do with it on my own. Because the libraries are so immense, I am going to give myself some leeway in terms of time to get through with it. I won't aim to do it all in a single day even though less that a dozen or so rewrites should suffice.

I will go at this like an immensely challenging problem. It is true that I looked over the libraries and have some idea of what their capabilities are, but tomorrow will be the first time using them."