"4:20pm. I am back. Let me read Kengan Omega and then I'll get a move on.
I've decided. Though it is theoretically possible to just study the libraries directly, it is really important to connect them to real world uses as well. Without that the motivation to keep going won't be possible to maintain. So going at it the way I wanted is just not practical. It would be a waste of time. The way I am doing it now has all the worst elements of school learning, and none of the good ones.

4:35pm. https://www.youtube.com/results?search_query=real+analysis&page=&utm_source=opensearch

There is a shitload on real analysis on Youtube as it turns out.

Very well. I'll go down this route. I guess I cannot avoid getting a general math education given what I want to do. I'll also make the reals chapter in the HoTT book one of my focuses. Let me just go through this normally.

4:45pm. https://youtu.be/sqEyWLGvvdw?list=PL0E754696F72137EC&t=253
"So Kroenecker had a very unusual point of view of mathematics. At least it was unusual at the time and it is in fact unusual today and that is...

He was what you might call a finitist. So this is a point of view that mathematics should only deal with finite objects, finite numbers, or things that could be constructed from the numbers in a finite number of steps."

4:55pm. https://youtu.be/sqEyWLGvvdw?list=PL0E754696F72137EC&t=738

I have to admit years and years ago I got some math books of the Internet. Among them there were some textbooks on real analysis. I cracked them open and it was just boring theorem after boring theorem. I dumped that shit pretty quickly.

Here he is doing a truly excellent job of setting up the subject and setting up the context for why would one even what to study this thing. I am definitely hooked.

6:05pm. https://youtu.be/gTgkrVATzmk?list=PL0E754696F72137EC

I'll stop here. I am done for the day.

I guess I should not be surprised that finite vs infinite has been an argument in math circles."