"10:10am. I went to bed on time, but woke up too early and I overslept when I decided to push myself to rest. Now I am so groggy I can barely think.
Well, I'll start the lectures soon.

10:40am. Let me do it.

11am. https://youtu.be/gTgkrVATzmk?list=PL0E754696F72137EC&t=1096

I personally can't do this in set theory because it is my first time seeing it, but I could prove this in Coq no problem.

11:55am. https://youtu.be/gTgkrVATzmk?list=PL0E754696F72137EC&t=3093

Hmmm, if p^2 is odd then p is odd?

12pm. https://youtu.be/gTgkrVATzmk?list=PL0E754696F72137EC&t=3389

This proof makes no sense to me, but whatever. Let me finish the lecture and I will take a break.

12:20pm. https://www.homeschoolmath.net/teaching/proof_square_root_2_irrational.php

Ah, I see it now. I am in a contrary mood and am trying to figure out the reason why the proof might be wrong.

Rather than proving that `sqrt 2` is irrational just because `a/b` is not simplified to lowest forms, isn't what this proof is saying that the simplification for `a/b` diverges? This might be a more useful way of looking at it.

Proofs that cannot ever possibly be computed should be invalid. Symbolic trickery is not real math.

Let me finish the lecture, only 20s more left.

https://youtu.be/fsMvCk9ZN7w?list=PL0E754696F72137EC

I'll stop here on lecture 3.

1:50pm. Let me resume. My capacity is really 3 lectures per day, so I have two more left in the day. Tomorrow I will take a break to do the review."