"2:30pm. https://youtu.be/fsMvCk9ZN7w?list=PL0E754696F72137EC&t=1220
The video quality for this is just so horrible. I am finding this unwatchable now that he is writing on the blackboard. Also these lectures are quite boring.

Let me find another series on this that is more palatable.

https://www.youtube.com/playlist?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu
Introductory Real Analysis

Let me go with this.

https://www.youtube.com/playlist?list=PL6763F57A61FE6FE8
Algebraic Topology

Since 2019 is looking like it will be shit, I might as well add algebraic topology to my list of things to study.

3pm. https://youtu.be/EaKLXK4hFFQ?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu&t=1038

Interesting that this course uses Mathematica. Since I got it a few months ago and have it taking up space, maybe I'll get a chance to make use of it. But to be honest, I just want to get through this course. I am not that deeply interested in this.

3:05pm. I can't watch this lecture. The lectures is one of those people who finds this kind of material engrossing rather than dull and he sort of expects the students to find it that way as well.

But since I am going to be studying HoTT, let me take a look at the algebraic topology lectures. After that I guess I'll get back to the lectures by the Asian guy.

3:15pm. https://youtu.be/Ap2c1dPyIVo?list=PL6763F57A61FE6FE8&t=455
"Q : What is algebraic topology?
A : Study of shapes and their properties independent of continuous deformation."

Yeah, this definitely does sound what I need to crack the credit assignment problem.

3:20pm. I guess I'll do this course a little and real analysis a little. I'll spread out the effort over the next few weeks and hopefully finish it by the end of September.

4:10pm. Done with the lecture. Let me finish the RA one that I dropped midway. I'll try to focus for a bit.

4:15pm. No screw the RA lectures. Let me check out the book by Rudin.

As expected, I had in on my drive from 5 years ago. I never though I would be willingly picking it up again.

4:50pm. 11/351. I am on like the second page of the book, and already I can't follow it.

5:05pm. Ok, I can follow it with some effort.

But damn this book goes quickly into it. It wastes no time unlike the lecture.

5:20pm. Maybe that is good.

"Consider the sets A and B from example 1.1. The set of A is bounded above. In fact the upper bounds of A are exactly the elements of B. Since B contains no smallest member, A has no least upper bound in Q."

The set `A : p^2 < 2` and `B : p^2 > 2`.

I understand what is meant by `B` not having a smallest member. I needed some reminding of what `A` and `B` are.

5:35pm. 14/351. I was not counting on something as tough as this at this point in the day.

I think I will stop here for the time being. I'll continue this tomorrow after I finish the review."