"2:30pm. https://www.nytimes.com/2006/05/07/books/chapters/0507-1st-gilb.html
Let me read this article. I guess I will take it easy today. After this excellent review I deserve a break.

This is one of the links I put in that essay. I first found the word nexting in a PhD by Adam White, but it was by this psychologist that it was introduced. I am not sure that I like the article so far.

Edit: It is fine.

3:35pm. Just a bit more and then I will start.

4:10pm. Let me start. I had a good rest. Thinking the above, I am not sure whether I made it obvious enough that path relinking cannot arise merely from equivalences, but that some additional condition is necessary, nevertheless, I hope I transmitted the feeling enough.

I believe in this.

What I do not believe though are real numbers. Which I should get to studying. Let me do it for an hour and then I am done for the day.

4:25pm. 14/351. I find the book by Rudin very off-putting. I do not like set theory.

I can barely musted up the focus to the read the stuff on this page.

https://pdfs.semanticscholar.org/8c95/95abdc915cc722ccfa30a7a0dc4566e11754.pdf
Constructive Real Analysis: a Type-Theoretical Formalization and Applications

This is a thesis by the author of the C-Corn library.

But forget that. I am going to try the lectures by the Indian and the woman first. Let me try the one by the later first.

I really wish the lectures by Stu were of better quality. But at 240p I can't see what is written on the board at all and this subject is boring enough that I need all the reinforcement that I can get.

https://youtu.be/uu4u_EYVuYw?t=84

This starts off with Lebesgue integrals and differentiation. This is fast.

https://www.youtube.com/playlist?list=PLcR4No3nvxFDg6VfaO8paSOv4ClAlyzon
Oh there is also this by Winston Ou.

https://www.youtube.com/playlist?list=PLLq_gUfXAnknTQofOxC3WpseyusORewKB
Here is the playlist by Eva Sinchich whose video I just linked.

https://www.youtube.com/playlist?list=PLXQazrgmw67EeQFcVW-oNoFBbtvGLUE3Y
Here is a playlist by William Grisaitis

https://www.youtube.com/playlist?list=PLbMVogVj5nJSxFihV-ec4A3z_FOGPRCo-
Here is the nptelhrd playlist by Kulkarni

https://www.youtube.com/playlist?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu
Here is the playlist by Bill Kinney that I also started yesterday.

I should just pick one and get on with it.

Hmmm...Let me just go with the later.

Rather than try few lectures a day which will make me burn out, I'll do one lecture per day and do something else as my main topic. I'll spread out my effort which should hopefully keep my interest going.

Ok, so I'll watch the lectures by Kinney. As my main topic it will be the PRML book. I'll use Agda + axioms to check my rewriting. Or maybe Coq. Forget about the libraries for a while. And maybe I'll do algebraic topology side by side as well. I'll try to get at least 2 items per day if I cannot 3.

4:55pm. I do have to add something to the review.

"I guess the overarching theme of this month's review is me realizing that data structures come first and algorithms second. You do need to invent integers first in order to be able to sort [Edit: No, you don't], and you do need reals before you can do differentiation. Maybe the reason ML is so difficult is because the primitives necessary for it have not been discovered yet."

Hmmm, no that would break the flow. It is good as it is. I do not feel like thinking about a proper sort analogy.

6pm. https://youtu.be/SpDQvVJZhYA?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu

Done with lunch and done with the first lecture for this. The first lecture was not too bad. I can deal with this.

Let me stop for the day here. I am looking forward to the next Symphogear ep, but till then let me play some Long War missions. I like how today went. I've been thinking how to do the review for days, and I am glad that all the pieces came together in a decent fashion."