"10:55am. https://youtu.be/SpDQvVJZhYA?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu&t=278
These arguments are such bullshit. Is this really worth studying? I can barely listen to this. I think I will just watch this lecture and then drop this nonsense.

11:05am. https://youtu.be/SpDQvVJZhYA?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu&t=697

Is this stuff not just brainwashing? HoTT did that to me a lot too, but it delivered in the end.

12:10pm. https://youtu.be/Crxv5TA5a6Y?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu

Let me stop here. In this lecture they will prove that sqrt 2 exists. I want to at least see how they do that, though no doubt it will be something stupid to disgust me again, but still spending an entire day on this is not too much of a price to pay.

I'll probably have to watch everything.

My goal with this is to understand enough to at least get me through the HoTT chapter. I know that indexed HITs are so I should have the knowledge necessary to understand how that stuff works. That will allow me to move on to algebraic topology. That seems like it could be promising."