"1pm. https://youtu.be/Crxv5TA5a6Y?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu&t=112
So that is how that proof went. This is breaking my mind. There is no way this is right. No way. This has to be some kind of paradox.

Ah, let me continue watching this.

1:15pm. Let me just say that I understand why they might think this is true. But the practical cause for denying the truth of that would be computability. There is also the logic that if `10` and `9` are not the same, and `10.0` and `9.9` are not the same, then why does repeating this pattern for an infinite amount suddenly make it true? You can literally see that it is not that same.

So he did show that it is true from one angle, but for something to be true it should be exhaustive.

1:30pm. https://en.wikipedia.org/wiki/0.999...

"The elementary argument of multiplying 0.333... = ​1⁄3 by 3 can convince reluctant students that 0.999... = 1. Still, when confronted with the conflict between their belief of the first equation and their disbelief of the second, some students either begin to disbelieve the first equation or simply become frustrated.[39] Nor are more sophisticated methods foolproof: students who are fully capable of applying rigorous definitions may still fall back on intuitive images when they are surprised by a result in advanced mathematics, including 0.999.... For example, one real analysis student was able to prove that 0.333... = ​1⁄3 using a supremum definition, but then insisted that 0.999... < 1 based on her earlier understanding of long division.[40] Others still are able to prove that ​1⁄3 = 0.333..., but, upon being confronted by the fractional proof, insist that "logic" supersedes the mathematical calculations.

Joseph Mazur tells the tale of an otherwise brilliant calculus student of his who "challenged almost everything I said in class but never questioned his calculator," and who had come to believe that nine digits are all one needs to do mathematics, including calculating the square root of 23. The student remained uncomfortable with a limiting argument that 9.99... = 10, calling it a 'wildly imagined infinite growing process.'"

This difference between `1/3` and `0.3...` is that the later pushes one into thinking of it like a value rather than a lazy computation which `1/3` would be.

It is like these educators are completely clueless that call-by-need and call-by-value are actual things in reality.

The reason why these proofs are abominable is because they pretend that the two things are the same, or that call-by-need is the only thing that exists.

No actually, it is not as simple as the calling convention. The underlying logic must be different as well.

1:45pm. What this issue bring to bear is that different things are valid depending on the underlying logic. It is not just a simple matter of saying that it is true or not like the lecturer seems to assume.

Ok. I'll accept that `0.999... = 1` might be true under certain (non-computable) logics.

1:50pm. https://youtu.be/Crxv5TA5a6Y?list=PLmU0FIlJY-MngWPhBDUPelVV3GhDw_mJu&t=615

Let me stop here. I do not feel like it anymore. I am through my budget for this abstract math stuff for the day.

Let me read the HoTT chapter. I'll chip away at it little by little.

2:25pm. 388/490. I can almost follow this, but I really lack too much background knowledge to properly grook this. Now that I understand the basics of cubical type theory and have a lens from which to interpret HoTT maybe it would be worth it to go through this book from the start.

Let me take a break here. I need it. I just do not feel like going at it full force. It seems that maintaining my focus for more than just one lecture is a challenge. This is how boring this stuff is. And I am not really interested in the logical paradoxes that the field of real analysis seems to be based on."