"3pm. It is not necessarily the case that `1/3 = 0.3~`.
To put it one way...

```
1/3 = sum {from=1; to=inf} (inl i -> 3 / 10^i)
```

I do not think there is any good way to actually prove that an infinite sum converges to a point. Because obviously this loop is going to go on forever.

You can only take it as an axiom that if God calculated it, that it would converge to that value. There is no other way. And axioms and computability do not mesh together. I am really wasting my time studying real analysis. What would really be amazing if they can actually prove that the above program converges to the value I expect despite being infinite. This mental trickery is dime and dozen in the wilds and holds no value whatsoever. Math - mind control. And not the good kind of mind control.

Let me resume the HoTT book.

3:20pm. 390/490. The stuff here so far is not enough for me to understand the Dedekind reals, but I am betting if I read that thesis I liked yesterday it should become more obvious.

Unlike when I looked at this chapter before, I can make out things, maybe due to my brief experience with the real analysis libraries.

4pm. 395/490. Why are the real numbers indexed by the rational ones?

4:25pm. This chapter is so damn huge, but I think if I could get it in computer code form that I could figure it out.

I just do not have the math talent necessary to ingest math in this verbal form. I need computer code to actually light the spark.

4:40pm. Ok, I sank some time into the chapter and now it is nearly 5pm. I have time for one more thing only. Let me try an algebraic topology lecture.

I think that I will try real analysis lectures for a few more days just to fully state my curiosity and then drop it. I'll pick it up again once Cubical Agda gets indexed HITs and constructive reals get added to the library.

4:40pm. https://youtu.be/oYFZaqArf54?list=PL6763F57A61FE6FE8&t=115
"I am not going to call it R^1. In other courses you would say this is the real number line, but I do not really believe in the real numbers. I believe that the rational numbers are the fundamental building block of mathematics and you've all been deluded and I've been deluded too many decades in think that the real numbers are the fundamental building block."

My brother!

But didn't this guy teach a course on differentiation? I remember watching the lectures last year during my information geometry phase.

Did he or did he not use real numbers there? I remember though he had a different method than usual for derivations. Something algebraic rather than using the limits.

...Egh, let me refresh my memory a little. I completely forgot about how to take derivatives using limits and it has been bugging me a while.

4:55pm. https://youtu.be/IePCHjMeFkE?t=498

Ah, I remember now. But I will leave the algebraic differentiation by NJ Wildberger for later.

5:20pm. https://youtu.be/oYFZaqArf54?list=PL6763F57A61FE6FE8&t=1344

Oh, this reasoning is exactly how it was done in HoTT.

10m left in this. Tomorrow I will definitely take some time to refresh my mind on his brand of differentiation.

5:30pm. https://www.youtube.com/watch?v=Ue5RTcRNGq0
Reorienting pure mathematics education and research | Insights into Mathematics | N J Wildberger

Let me watch this for a while. I am done with studying for the day.

"I have quite non-standard view about the direction of not just pure mathematics research, but also and more importantly mathematics education. And those two things are intimately connected from my point of view. In this channel we are going to be exploring new directions for mathematics. Both mathematics education and research. By concentrating our efforts towards foundationally explicit and correct underpinnings for the subject. We'll be looking at the foundations of mathematics in much more detail and with much more care than any other place does. So I have serious reservations about the set theoretical foundations of modern mathematics. I do not believe in infinite processes, in the sense that I do not believe that we are able logically to obtain the result of an infinite process."

This is exactly why I dislike real analysis so heavily. It is just what I wrote in the previous entires. I guess this guy is worth listening to.

"We might talk about the infinite process as a computer program that does not end, but it is wrong for us to pretend that we can get the output of such a process and use it as ingredients for further mathematical developments. And that is in essence what a lot of mathematics is currently doing.

So a lot of analysis, a lot of topology, a lot of even algebraic geometry and other subject are one way or the other building in this kind of assumption that we can get to the end of infinity, extract something and take that something back into mathematics that we are doing.

This is logically unjustified and I am looking for alternative ways."

5:50pm. I guess I'll check out Algebraic Calculus tomorrow.

https://www.youtube.com/channel/UCriFv3G22iOUidUhkIGXuhw
Wild Egg mathematics courses

https://www.youtube.com/watch?v=cINtOxgDWNc&list=PLIljB45xT85BshlZ5HaiINGugNOKeGpw8
Algebraic Calculus One - Now under way!

Ah, let me watch this for a bit since I already started it.

6:50pm. I ended up watching that entire last video. Let me call it a day here."