"11:40am. No forget this. I think I'll just drop this and move to Wildberger's Algebraic Calculus course instead.
12:30pm. Done with breakfast. Let me finally start

12:55pm. https://youtu.be/rTw6XbmO8Nc?t=1119

So it will be a paid course. Well, I'll take what I can get for free.

I already have Maple and Mathematica. Maybe I will get GeoGebra.

https://www.youtube.com/watch?v=XQH4bb0SIls&list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&index=3&t=0s
Points and Lines in the Affine Plane | Algebraic Calculus One | Wild Egg

I might as well watch this. So far, my quest has been completely derailed. I should be reading the PRML book at this point, but what the hell.

1:55pm. Let me take a break here. The lecture is boring me to death as it is just about lines and points.

As the last thing in real analysis until Cubical Agda gets HITs, I'll go for for that Corn paper (Constructive Real Analysis).

3:35pm. https://pdfs.semanticscholar.org/8c95/95abdc915cc722ccfa30a7a0dc4566e11754.pdf

"As explained in Appendix A, there is a preferred sort for propositions in Coq, namely Prop, whereas datatypes usually have type Set. One of the main purposes of this distinction, which will be explained in more detail in Chapter 6, is to separate the computational content of proofs from the pure logical reasoning. In particular, data is not allowed to depend on propositions; in other words, if A : Prop and S : Set, then case analysis on elements of type A is not allowed when defining an element of type S.

But this is precisely what mathematicians do all the time! Defining functions by cases (i.e., analyzing the structure of a term of type A ∨ B) is a daily task when proving theorems in Analysis, and this is precisely one of the things that the previous rule forbids—even though it is clearly justified in constructive mathematics"

Yeah, this is one of the things that bugs me about Coq and that Agda fixes.

"Providing a good documentation for the formalized library in parallel with its development was a central preoccupation since the early stages of the FTA-project. Having a human-readable description of what has been formalized can be very useful in communicating not only content but also ideas, notations and even some technical aspects of the formalization process."

Maybe it might be worth giving this library a chance. Coquelicot has zero in terms of documentation.

4pm. This paper is such a drag. That is it. I am done with real analysis. So far every subject I've ever looked at on this just drones on and on and on. Holy shit.

I am sticking to algebra. I'll make do with just it.

It is algebraic topology and algebraic calculus from here on out for me.

4:20pm. I'll admit I am really interested in how you can do differentiation with just rationals. I keep writing that differentiation comes from the reals, even though the Wildberger repeatedly stated that rationals will suffice.

It would just be too much of a pain right now to adjust my story. I'll do it when I actually understand the new theory.

Now let me watch his lectures. I feel like I am just flushing time down the drain. Living like this is just sad."