"4:30pm. I need to start, right now I've just been replaying things in my mind.
I dropped the lecture I was watching mid-way so let me wrap it up. At the very least, lines and points is something I can grok unlike real analysis.

5:10pm. https://youtu.be/JJVtfeto-AY?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=247

Why is the point written like a proportion with square brackets? It would make more sense to write it as a product.

5:35pm. https://youtu.be/JJVtfeto-AY?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=662

The projective test formula is not at all obvious here. The way I'd check it would be to normalize the both sides by the third field, but what I see here would require some rewriting to arrive at. Interesting. I guess I'll learn a thing or two during this course.

5:55pm. https://youtu.be/tIQU3H3K7_Q?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

I'll stop on this lecture. I remember now that the formula is the cross product. I am not too interested in the reasoning behind it.

These lectures are not too tough, so I should be able to go through quite a few in a day once I put my mind to it.

6pm. The plan for the next few weeks will be to just watch lectures by Wildberger. Then once I get back to PRML, I'll use Agda or Coq axiomatically to double check my rewriting. Considering how much effort I put into learning theorem proving, it seems disappointing that this is what I will be doing, but I do feel like I've gained quite a bit of understanding from all the exercises. Even though it is not the understanding that I'd have preferred, since I now have it I might as well accept it.

Hmmm...

https://github.com/agda/agda-stdlib/blob/master/src/Data/Rational.agda

In place of reals, I suppose I could make do with rationals. It is just that I do need division and would rather not make it an axiom on the integers.

6:05pm. Let me take off here. It is really sad, that it is September 2019 and I still haven't gotten any closer to making those agents. Money eludes me. I just can't have anything go the way I want.

I really do seem to be destined to pay my respect to the efforts the old greats put in. Surely they did not do their pioneering work so the world dies in darkness. They surely have not done their research so a bunch of lecturers can gape at how smart they were as they bore their audience to death. They did not do their work so they could fill out a library.

I won't accept that kind of history.

Math isn't supposed to be about solving an endless series of meaningless puzzles. I will prove it. And in doing so, create the light."